Per-batch cost edits

+++ contracts/src/precompiles/ArbGasInfo.sol
@@ -114,7 +114,7 @@ interface ArbGasInfo {
     function getL1PricingSurplus() external view returns (int256);
 
     /// @notice Returns the base charge (in L1 gas) attributed to each data batch in the calldata pricer
-    function getPerBatchGasCharge() external view returns (int256);
+    function getPerBatchGasCharge() external view returns (int64);
 
     /// @notice Returns the cost amortization cap in basis points
     function getAmortizedCostCapBips() external view returns (uint64);
